# openapi.yaml
openapi: '3.0.3'
info:
  title: Career-Copilot API
  description: AI模拟面试官与智能职业规划平台 - 后端接口文档
  version: '1.0.0'
  contact:
    name: 陶宏阳
    url: https://github.com/peeker-tao
servers:
  - url: http://localhost:3002/api
    description: 本地开发环境
  - url: https://test-api.career-copilot.com/api/v1
    description: 测试环境
  - url: https://api.career-copilot.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 Authorization 头中传入 Bearer Token
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP 状态码
          example: 200
        message:
          type: string
          description: 提示信息
          example: success
        data:
          type: object
          description: 业务数据
      required:
        - code
        - message
        - data
...
    PaginationResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP 状态码
          example: 400
        message:
          type: string
          description: 错误提示
          example: 请求参数错误
        data:
          type: object
          description: 错误详情（可选）
      required:
        - code
        - message

    # ============================================================
    # 认证模块
    # ============================================================
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          description: 邮箱，需合法格式
          example: user@example.com
        password:
          type: string
          minLength: 6
          description: 密码，不少于 6 位
          example: abc123456
        name:
          type: string
          minLength: 2
          maxLength: 20
          description: 昵称，2-20 字符
          example: 小明
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: 注册邮箱
          example: user@example.com
        password:
          type: string
          description: 密码
          example: abc123456
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: 刷新令牌
          example: eyJhbG...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          description: 用户 ID
          example: clx...
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: 小明
        avatar:
          type: string
          format: uri
          nullable: true
          example: https://...
        education:
          type: string
          nullable: true
          description: 教育背景
          example: 华中科技大学 软件工程 本科 2024级
        targetPosition:
          type: string
          nullable: true
          description: 目标岗位
          example: 后端开发工程师
        role:
          type: string
          enum:
            - user
            - admin
          description: 用户角色
          example: user
        createdAt:
          type: string
          format: date-time
          example: '2026-06-13T08:00:00.000Z'
    AuthResult:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserInfo'
        accessToken:
          type: string
          description: 访问令牌
          example: eyJhbG...
        refreshToken:
          type: string
          description: 刷新令牌
          example: eyJhbG...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 20
          description: 昵称
          example: 小明改名字了
        avatar:
          type: string
          format: uri
          description: 头像 URL
          example: https://...
        education:
          type: string
          description: 教育背景
          example: 武汉大学 计算机科学与技术 硕士 2024级
        targetPosition:
          type: string
          description: 目标岗位
          example: 算法工程师

    # ============================================================
    # 简历模块
    # ============================================================
...
    CreateInterviewRequest:
      type: object
      required:
        - targetPosition
        - difficulty
      properties:
        resumeId:
          type: string
          description: 关联简历 ID（可选，传了则基于简历出题）
          example: clx...
        targetPosition:
          type: string
          description: 目标岗位
          example: 后端开发工程师
        difficulty:
          type: string
          enum:
            - easy
            - medium
            - hard
          description: 面试难度
          example: medium
    InterviewSummary:
      type: object
      properties:
        id:
          type: string
          example: intv_xxx
        targetPosition:
          type: string
          example: 后端开发工程师
        difficulty:
          type: string
          enum:
            - easy
            - medium
            - hard
          example: medium
        status:
          type: string
          enum:
            - in_progress
            - completed
          example: completed
        score:
          type: integer
          nullable: true
          example: 78
        questionCount:
          type: integer
          example: 8
        startedAt:
          type: string
          format: date-time
          example: '2026-06-13T08:00:00.000Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-06-13T08:25:00.000Z'
    InterviewDetail:
      allOf:
        - $ref: '#/components/schemas/InterviewSummary'
    CreateInterviewResult:
      allOf:
        - $ref: '#/components/schemas/InterviewSummary'
        - type: object
          properties:
            firstQuestion:
              type: string
              description: AI 第一个面试题
              example: 请介绍一下 Java 中的 HashMap 底层实现原理？
    InterviewMessage:
      type: object
      properties:
        id:
          type: string
          example: msg_1
        role:
          type: string
          enum:
            - ai
            - user
          description: 消息发送者角色
          example: ai
        content:
          type: string
          description: 消息内容
          example: 请介绍一下 Java 中 HashMap 的底层实现原理？
        questionType:
          type: string
          nullable: true
          enum:
            - technical
            - behavioral
            - project
          description: 题目类型
          example: technical
        feedback:
          type: string
          nullable: true
          description: AI 对回答的即时反馈
          example: '评分：4/5。对底层结构描述清晰，但扩容机制遗漏。'
        createdAt:
          type: string
          format: date-time
          example: '2026-06-13T08:00:00.000Z'
    SubmitAnswerRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: 用户回答内容
          example: HashMap 底层是数组+链表+红黑树实现的...
    AiResponse:
      type: object
      properties:
        id:
          type: string
          example: msg_3
        role:
          type: string
          enum:
            - ai
          example: ai
        content:
          type: string
          description: AI 的回应内容
          example: 回答得不错，但我注意到你没有提到负载因子和扩容机制，能否再详细说明一下？
        feedback:
          type: string
          description: AI 评分与反馈
          example: '评分：4/5。对底层结构描述清晰，但扩容机制遗漏。'
        isFollowUp:
          type: boolean
          description: 是否追问（true=继续追问，false=进入下一题）
          example: true
        questionType:
          type: string
          enum:
            - technical
            - behavioral
            - project
          example: technical
    SubmitAnswerResult:
      type: object
      properties:
        userMessageId:
          type: string
          example: msg_2
        aiResponse:
          $ref: '#/components/schemas/AiResponse'
    AnswerDetail:
      type: object
      properties:
        question:
          type: string
          description: 面试问题
          example: HashMap 实现原理
        score:
          type: integer
          description: 该题得分（1-5）
          example: 4
        comment:
          type: string
          description: 评语
          example: 底层结构答得好，扩容机制遗漏
    Suggestion:
      type: object
      properties:
        item:
          type: string
          description: 改进建议
          example: 深入学习 Redis、消息队列等中间件
        priority:
          type: string
          enum:
            - high
            - medium
            - low
          example: high
        resources:
          type: array
          items:
            type: string
          example:
            - '《Redis 设计与实现》'
            - MIT 6.824 分布式系统
    InterviewFeedback:
      type: object
      properties:
        interviewId:
          type: string
          example: intv_xxx
        overallScore:
          type: integer
          description: 综合评分（百分制）
          example: 78
        dimensions:
          type: object
          description: 各维度评分
          additionalProperties:
            type: object
            properties:
              score:
                type: integer
                example: 80
              comment:
                type: string
                example: 专业技能扎实，但对分布式相关知识了解不足
          example:
            professionalSkill:
              score: 80
              comment: 专业技能扎实，但对分布式相关知识了解不足
            communication:
              score: 75
              comment: 表达较清晰，但结构可以更严谨
            logicalThinking:
              score: 82
              comment: 逻辑思维能力强
            projectExperience:
              score: 70
              comment: 项目介绍缺乏亮点和数据支撑
        strengths:
          type: array
          items:
            type: string
          example:
            - Java 基础扎实
            - 逻辑清晰
            - 学习态度积极
        weaknesses:
          type: array
          items:
            type: string
          example:
            - 分布式经验不足
            - 项目深度不够
            - 缺乏系统设计思维
        improvementSuggestions:
          type: array
          items:
            $ref: '#/components/schemas/Suggestion'
        answerDetails:
          type: array
          items:
            $ref: '#/components/schemas/AnswerDetail'

    # ============================================================
    # 职业规划模块
    # ============================================================
...
paths:
...
  /auth/register:
    post:
      tags:
        - 认证管理
      summary: 用户注册
      description: 使用邮箱和密码注册新用户，注册成功自动返回 Token
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 201
                      message:
                        type: string
                        example: 注册成功
                      data:
                        $ref: '#/components/schemas/AuthResult'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 邮箱已注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - 认证管理
      summary: 用户登录
      description: 使用邮箱和密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResult'
        '401':
          description: 邮箱或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - 认证管理
      summary: 刷新 Token
      description: 使用 Refresh Token 获取新的 Access Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                            example: eyJhbG...
                          refreshToken:
                            type: string
                            example: eyJhbG...
        '401':
          description: Refresh Token 无效或过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    get:
      tags:
        - 认证管理
      summary: 获取用户信息
      description: 获取当前登录用户的详细信息
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
        '401':
          description: 未认证或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - 认证管理
      summary: 修改个人资料
      description: 修改当前登录用户的个人资料
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfo'
        '401':
          description: 未认证或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # 简历模块
  # ============================================================
...
  /interviews:
    post:
      tags:
        - 面试管理
      summary: 创建面试会话
      description: 创建一个新的 AI 模拟面试会话
      operationId: createInterview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInterviewRequest'
      responses:
        '201':
          description: 面试会话创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 201
                      message:
                        type: string
                        example: 面试会话创建成功
                      data:
                        $ref: '#/components/schemas/CreateInterviewResult'
        '400':
          description: 参数错误
    get:
      tags:
        - 面试管理
      summary: 获取面试历史
      description: 获取当前用户的所有面试记录，支持筛选
      operationId: listInterviews
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
          description: 每页条数
        - name: status
          in: query
          schema:
            type: string
            enum:
              - in_progress
              - completed
          description: 按状态筛选
        - name: targetPosition
          in: query
          schema:
            type: string
          description: 按目标岗位筛选
      responses:
        '200':
          description: 成功获取面试历史
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/InterviewSummary'
                          pagination:
                            $ref: '#/components/schemas/PaginationResult'
  /interviews/{id}:
    get:
      tags:
        - 面试管理
      summary: 获取面试详情
      description: 获取单次面试的基本信息
      operationId: getInterview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 面试会话 ID
          example: intv_xxx
      responses:
        '200':
          description: 成功获取面试详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InterviewDetail'
        '404':
          description: 面试会话不存在
  /interviews/{id}/messages:
    get:
      tags:
        - 面试管理
      summary: 获取面试对话记录
      description: 获取指定面试会话的完整对话消息列表
      operationId: getInterviewMessages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 面试会话 ID
      responses:
        '200':
          description: 成功获取对话记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          messages:
                            type: array
                            items:
                              $ref: '#/components/schemas/InterviewMessage'
        '404':
          description: 面试会话不存在
  /interviews/{id}/answer:
    post:
      tags:
        - 面试管理
      summary: 提交回答
      description: 提交用户对当前面试题的回答，AI 将给出反馈并决定是否追问
      operationId: submitAnswer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 面试会话 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: 成功提交回答，返回 AI 回应
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmitAnswerResult'
        '400':
          description: 面试已结束或参数错误
        '404':
          description: 面试会话不存在
  /interviews/{id}/feedback:
    post:
      tags:
        - 面试管理
      summary: 获取面试报告
      description: 面试结束后，生成并获取详细的面试评分与改进建议报告
      operationId: getInterviewFeedback
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 面试会话 ID（状态须为 completed）
      responses:
        '200':
          description: 成功获取面试报告
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InterviewFeedback'
        '400':
          description: 面试尚未结束
        '404':
          description: 面试会话不存在

  # ============================================================
  # 职业规划模块
  # ============================================================
...